one_gadget as library, closed #47
